Add some benchmarks akin to philipc/dwarf-bench
The original raw-ffi benchmark is included, tweaked only to fix
differences in crate paths.

There are two versions using the new abstractions, and the difference is
only how they walk die children.  The first uses `children` as a normal
`Iterator`, and it's a fair amount slower due to many mem copies.  The
second uses `with_children` for directly nested calls, and this actually
*beats* the raw-ffi version.  I have new respect for optimizing copies.

It's probably a fair followup to try optimizing the ffi version too,
giving it a recursive style instead of the manual "stack", for one.